move cart editing methods out of main Order file.
